fix(crypto): continue refactor of imports and instantiations to resolve mypy errors
