+++ cashu/core/crypto/keys.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Optional, Union
 from bip32 import BIP32
 
 from .bls import PrivateKey as BlsPrivateKey, PublicKey as BlsPublicKey
-from .interfaces import ICashuPrivateKey, ICashuPublicKey, PrivateKey, PublicKey
+from .interfaces import ICashuPrivateKey, ICashuPublicKey
 from .secp import SecpPrivateKey, SecpPublicKey
 
 
@@ -40,7 +40,7 @@ def derive_keys_deprecated_pre_0_15(
     }
 
 
-def derive_pubkey(seed: str) -> PublicKey:
+def derive_pubkey(seed: str) -> SecpPublicKey:
     pubkey = SecpPrivateKey(
         hashlib.sha256((seed).encode("utf-8")).digest()[:32],
     ).public_key
@@ -54,7 +54,7 @@ def derive_pubkeys(
     return {amt: keys[amt].public_key for amt in amounts}
 
 
-def derive_keyset_id(keys: Dict[int, PublicKey]):
+def derive_keyset_id(keys: Dict[int, Union[SecpPublicKey, BlsPublicKey]]):
     """Deterministic derivation keyset_id from set of public keys (version 00)."""
     # sort public keys by amount
     sorted_keys = dict(sorted(keys.items()))
@@ -63,7 +63,7 @@ def derive_keyset_id(keys: Dict[int, PublicKey]):
 
 
 def derive_keyset_id_v2(
-    keys: Dict[int, PublicKey], 
+    keys: Dict[int, Union[SecpPublicKey, BlsPublicKey]], 
     unit: str, 
     final_expiry: Optional[int] = None,
     input_fee_ppk: int = 0,
@@ -101,6 +101,7 @@ def derive_keyset_id_v2(
     
     # prefix with version byte 01
     return f"01{hash_digest}"
+
 
 
 def derive_keyset_short_id(keyset_id: str) -> str:
@@ -190,7 +191,7 @@ def is_keyset_id_v2(keyset_id: str) -> bool:
     return get_keyset_id_version(keyset_id) == '01'
 
 
-def derive_keyset_id_deprecated(keys: Dict[int, PublicKey]):
+def derive_keyset_id_deprecated(keys: Dict[int, Union[SecpPublicKey, BlsPublicKey]]):
     """DEPRECATED 0.15.0: Deterministic derivation keyset_id from set of public keys.
     DEPRECATION: This method produces base64 keyset ids. Use `derive_keyset_id` instead.
     """
@@ -202,29 +203,8 @@ def derive_keyset_id_deprecated(keys: Dict[int, PublicKey]):
     ).decode()[:12]
 
 
-def random_hash() -> str:
-    """Returns a base64-urlsafe encoded random hash."""
-    return base64.urlsafe_b64encode(
-        bytes([random.getrandbits(8) for i in range(30)])
-    ).decode()
-
-def derive_keys_v3(mnemonic: str, derivation_path: str, amounts: List[int]) -> Dict[int, BlsPrivateKey]:
-    """
-    Deterministic derivation of BLS12-381 keys for 2^n values.
-    Since BIP32 doesn't technically cover BLS12-381, we use HKDF or simple hashing on the BIP32 seed.
-    For simplicity and backwards compatibility of mnemonic/path logic, we hash the BIP32 path output to generate the scalar.
-    """
-    bip32 = BIP32.from_seed(mnemonic.encode())
-    orders_str = [f"/{a}'" for a in range(len(amounts))]
-    return {
-        a: BlsPrivateKey(
-            hashlib.sha256(bip32.get_privkey_from_path(derivation_path + orders_str[i])).digest()
-        )
-        for i, a in enumerate(amounts)
-    }
-
 def derive_keyset_id_v3(
-    keys: Dict[int, PublicKey], 
+    keys: Dict[int, Union[SecpPublicKey, BlsPublicKey]], 
     unit: str, 
     final_expiry: Optional[int] = None,
     input_fee_ppk: int = 0,
@@ -241,3 +221,11 @@ def derive_keyset_id_v3(
         keyset_id_bytes += f"|final_expiry:{final_expiry}".encode("utf-8")
     hash_digest = hashlib.sha256(keyset_id_bytes).hexdigest()
     return f"02{hash_digest}"
+
+
+def random_hash() -> str:
+    """Returns a base64-urlsafe encoded random hash."""
+    return base64.urlsafe_b64encode(
+        bytes([random.getrandbits(8) for i in range(30)])
+    ).decode()
+
