+++ cashu/wallet/auth/auth.py
@@ -7,9 +7,10 @@ from loguru import logger
 from cashu.core.helpers import sum_proofs
 from cashu.core.mint_info import MintInfo
 
-from ...core.base import AnyPrivateKey, Proof
-from ...core.crypto.secp import PrivateKey as SecpPrivateKey
-from ...core.db import Database
+from cashu.core.base import Proof
+from cashu.core.crypto.interfaces import PrivateKey, PublicKey
+from cashu.core.crypto.secp import SecpPrivateKey, SecpPublicKey
+from cashu.core.db import Database
 from ..crud import get_mint_by_url, update_mint
 from ..wallet import Wallet
 from .openid_connect.openid_client import AuthorizationFlow, OpenIDClient
